+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## 2.11.0
 
 * Added an optional `postage` argument to `post_precompiled_letter` method and `send_precompiled_letter` that calls it.
-* Added postage to `ResponseNotification` and to `Notification` class on the client.
+* Added postage to `Notification` class on the client.
+* Added new response class: `ResponsePrecompiledLetter`, which makes response object for precompiled letters different
+class to other notification response objects.
 
 ## 2.10.0